Using formatter that shows defaults automatically
The `ArgumentDefaultsHelpFormatter` automatically shows the defaults. This
ensures two things:

- The db user/pass/etc defaults are shown (they were previously missing, which
  I needed for setting up travis :-) )

- The defaults always match exactly the actual defaults. There is no
  possibility for the documentation and the code to drift apart.

I had to remove `metavar` in some places, as that one gets used instead of
defaults if it is specified.

Note that the extra message about some options not being implemented isn't
shown anymore due to the formatting changes. If you want that back, I'd
suggest an addition to `ArgumentParser()`: `epilog="Note, x and y aren't
implemented yet"`